+++ cmd/service/service.go
@@ -13,7 +13,7 @@ const githubURL = "https://api.github.com"
 //Client is interface of service
 type Client interface {
 	GetUser() (types.UserInfo, error)
-	GetFollowing() (types.FollowingUserList, error)
+	GetFollowing() (types.FollowingUsers, error)
 	GetFollowers() (types.Followers, error)
 }
 
@@ -63,14 +63,14 @@ func (config config) GetUser() (types.UserInfo, error) {
 }
 
 //GetFollowing get following user information from github.com
-func (config config) GetFollowing() (types.FollowingUserList, error) {
+func (config config) GetFollowing() (types.FollowingUsers, error) {
 	resp, err := makeRequest(http.MethodGet, config.URL, nil)
 	if err != nil {
-		return types.FollowingUserList{}, err
+		return types.FollowingUsers{}, err
 	}
-	userInfoList := make(types.FollowingUserList, 0)
+	userInfoList := make(types.FollowingUsers, 0)
 	if err := json.NewDecoder(resp.Body).Decode(&userInfoList); err != nil {
-		return types.FollowingUserList{}, fmt.Errorf("error decoding response %v", err)
+		return types.FollowingUsers{}, fmt.Errorf("error decoding response %v", err)
 	}
 
 	return userInfoList, nil
